refactor: rename exception name
