+++ src/main/java/nextstep/subway/advice/LineControllerAdvice.java
@@ -1,7 +1,6 @@
 package nextstep.subway.advice;
 
-import nextstep.subway.exception.DuplicateStationException;
-import nextstep.subway.exception.UnmatchedLastStationAndNewUpStationException;
+import nextstep.subway.exception.AddSectionException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -9,13 +8,8 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 @RestControllerAdvice
 public class LineControllerAdvice {
 
-    @ExceptionHandler(UnmatchedLastStationAndNewUpStationException.class)
-    public ResponseEntity<ErrorResult> unmatchedLastStationAndNewUpStationException(UnmatchedLastStationAndNewUpStationException e) {
-        return ResponseEntity.badRequest().body(new ErrorResult(e.getMessage()));
-    }
-
-    @ExceptionHandler(DuplicateStationException.class)
-    public ResponseEntity<ErrorResult> duplicateStationException(DuplicateStationException e) {
+    @ExceptionHandler(AddSectionException.class)
+    public ResponseEntity<ErrorResult> duplicateStationException(AddSectionException e) {
         return ResponseEntity.badRequest().body(new ErrorResult(e.getMessage()));
     }
 }
