--- src/main/java/nextstep/subway/exception/UnmatchedLastStationAndNewUpStationException.java
@@ -1,8 +1,0 @@
-package nextstep.subway.exception;
-
-public class UnmatchedLastStationAndNewUpStationException extends RuntimeException {
-
-    public UnmatchedLastStationAndNewUpStationException(String message) {
-        super(message);
-    }
-}
